Adding bg option to watershed
This option replaces the hardcoded BG macro at the C level, giving users
flexibility on the pixel value threshold to apply when finding pixels.
This reduces the number of pixels to be inspected in images with low
background noises. Tolerance is still defined in absolute terms.

Note that we use the bg value not only to filter out which pixels are to
be considered, but also to populate the frame array. This is important
as it gives a more stable set of values to the sorting routine, yielding
improved speed.

I also took the opportunity of correctly aligning a couple of lines of
code that had tabs in them and therefore were a bit off.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>